+++ packages/editor/editor-core/src/__tests__/visual-regression/z-indexes.ts
@@ -25,7 +25,8 @@ const mentionButton = `span[aria-label="${
 }"]`;
 const mentionQuery = 'span[data-type-ahead-query]';
 
-describe('Snapshot Test: z-indexes', () => {
+// TODO - add ADF before loading stuff
+describe.skip('Snapshot Test: z-indexes', () => {
   ['full-page', 'comment'].forEach(appearance => {
     let page;