+++ packages/elements/user-picker/src/components/UserPicker.tsx
@@ -39,10 +39,9 @@ import {
   optionToSelectableOptions,
 } from './utils';
 
-class UserPickerInternal extends React.Component<
-  UserPickerProps & WithAnalyticsEventProps,
-  UserPickerState
-> {
+type Props = UserPickerProps & WithAnalyticsEventProps;
+
+class UserPickerInternal extends React.Component<Props, UserPickerState> {
   static defaultProps: UserPickerProps = {
     width: 350,
     isMulti: false,
@@ -77,11 +76,11 @@ class UserPickerInternal extends React.Component<
     return derivedState;
   }
 
-  private selectRef;
+  private selectRef: any | null;
 
   private session?: UserPickerSession;
 
-  constructor(props) {
+  constructor(props: Props) {
     super(props);
     this.state = {
       options: [],
@@ -157,7 +156,7 @@ class UserPickerInternal extends React.Component<
     }
   };
 
-  private handleSelectRef = ref => {
+  private handleSelectRef = (ref: any | null) => {
     this.selectRef = ref;
   };
 
@@ -168,7 +167,10 @@ class UserPickerInternal extends React.Component<
           return {
             options: options.concat(
               newOptions.reduce<OptionData[]>(
-                (nextOptions, item) => nextOptions.concat(item[0]),
+                (nextOptions, item) =>
+                  Array.isArray(item)
+                    ? nextOptions.concat(item[0])
+                    : nextOptions.concat(item),
                 [],
               ),
             ),
@@ -193,7 +195,7 @@ class UserPickerInternal extends React.Component<
         const addOptions = this.addOptions.bind(
           this,
           inflightRequest.toString(),
-        );
+        ) as (value: OptionData | OptionData[]) => void | PromiseLike<void>;
         let count = 0;
         if (isIterable(result)) {
           for (const value of result) {
@@ -245,6 +247,7 @@ class UserPickerInternal extends React.Component<
     { action }: { action: InputActionTypes },
   ) => {
     if (action === 'input-change') {
+      callCallback(this.props.onInputChange, search);
       this.setState({ inputValue: search, preventFilter: false });
 
       this.executeLoadOptions(search);
@@ -265,7 +268,7 @@ class UserPickerInternal extends React.Component<
     this.fireEvent(focusEvent);
   };
 
-  componentDidUpdate(prevProps: UserPickerProps, prevState: UserPickerState) {
+  componentDidUpdate(_: UserPickerProps, prevState: UserPickerState) {
     const { menuIsOpen, options } = this.state;
     // load options when the picker open
     if (menuIsOpen && !prevState.menuIsOpen) {
@@ -293,7 +296,6 @@ class UserPickerInternal extends React.Component<
       if (this.session) {
         this.session.inputChangeTime = Date.now();
       }
-      callCallback(this.props.onInputChange, this.state.inputValue);
     }
   }
 
@@ -342,7 +344,6 @@ class UserPickerInternal extends React.Component<
 
   render() {
     const {
-      width,
       isMulti,
       anchor,
       isLoading,
@@ -363,6 +364,7 @@ class UserPickerInternal extends React.Component<
       inputValue,
     } = this.state;
     const appearance = this.getAppearance();
+    const width = this.props.width as string | number;
 
     return (
       <Select
@@ -372,7 +374,7 @@ class UserPickerInternal extends React.Component<
         isMulti={isMulti}
         options={this.getOptions()}
         onChange={this.handleChange}
-        styles={getStyles(width, appearance)}
+        styles={getStyles(width)}
         components={getComponents(isMulti, anchor)}
         inputValue={inputValue}
         menuIsOpen={menuIsOpen}