+++ packages/media/media-avatar-picker/src/index.tsx
@@ -1,2 +1,4 @@
-export * from './avatar-picker-dialog';
+export {
+  default as AvatarPickerDialog,
+} from './avatar-picker-dialog/avatar-picker-dialog-loader';
 export { Avatar } from './avatar-list';