+++ packages/navigation/centralised-app-switcher/src/components/confluence-app-switcher.tsx
@@ -4,6 +4,7 @@ import {
   AppSwitcherItem,
   Section,
   ManageButton,
+  Skeleton,
 } from '../primitives';
 import { CustomLinksProvider } from '../providers/confluence-data-providers';
 import {
@@ -38,8 +39,7 @@ export default ({ cloudId }: WithCloudId) => {
                   }) => (
                     <AppSwitcherWrapper>
                       {isLoadingRecentContainers ? (
-                        // TODO: Add proper skeleton component once it's ready https://hello.atlassian.net/browse/CEN-47
-                        'Loading Recent Containers...'
+                        <Skeleton />
                       ) : (
                         <Section title="Recent Containers">
                           {recentContainersData &&
@@ -53,8 +53,7 @@ export default ({ cloudId }: WithCloudId) => {
                         </Section>
                       )}
                       {isLoadingCustomLinks ? (
-                        // TODO: Add proper skeleton component once it's ready https://hello.atlassian.net/browse/CEN-47
-                        'Loading Custom Links...'
+                        <Skeleton />
                       ) : (
                         <Section title="Custom Links">
                           {customLinksData &&
@@ -68,8 +67,7 @@ export default ({ cloudId }: WithCloudId) => {
                         </Section>
                       )}
                       {isLoadingLicenseInformation ? (
-                        // TODO: Add proper skeleton component once it's ready https://hello.atlassian.net/browse/CEN-47
-                        'Loading License Information...'
+                        <Skeleton />
                       ) : (
                         <Section title="License Information">
                           {licenseInformationData &&
@@ -86,8 +84,7 @@ export default ({ cloudId }: WithCloudId) => {
                         </Section>
                       )}
                       {isLoadingAdminPermission ? (
-                        // TODO: Add proper skeleton component once it's ready https://hello.atlassian.net/browse/CEN-47
-                        'Loading Admin Permission...'
+                        <Skeleton />
                       ) : (
                         <Section title="Admin Permission">
                           {adminPermissionData && (