+++ packages/navigation/centralised-app-switcher/src/components/jira-app-switcher.tsx
@@ -4,6 +4,7 @@ import {
   AppSwitcherItem,
   Section,
   ManageButton,
+  Skeleton,
 } from '../primitives';
 import { CustomLinksProvider } from '../providers/jira-data-providers';
 import {
@@ -82,8 +83,7 @@ export default ({ cloudId, triggerXFlow }: JiraAppSwitcherProps) => {
                         )}
                       {isLoadingLicenseInformation ||
                       isLoadingAdminPermission ? (
-                        // TODO: Add proper skeleton component once it's ready https://hello.atlassian.net/browse/CEN-47
-                        'Loading License Information...'
+                        <Skeleton />
                       ) : (
                         <Section title="Products">
                           {licenseInformationData &&
@@ -130,8 +130,7 @@ export default ({ cloudId, triggerXFlow }: JiraAppSwitcherProps) => {
                         </Section>
                       )}
                       {isLoadingCustomLinks ? (
-                        // TODO: Add proper skeleton component once it's ready https://hello.atlassian.net/browse/CEN-47
-                        'Loading Custom Links...'
+                        <Skeleton />
                       ) : (
                         <Section title="More" isCustom>
                           {customLinksData &&
@@ -145,8 +144,7 @@ export default ({ cloudId, triggerXFlow }: JiraAppSwitcherProps) => {
                         </Section>
                       )}
                       {isLoadingRecentContainers ? (
-                        // TODO: Add proper skeleton component once it's ready https://hello.atlassian.net/browse/CEN-47
-                        'Loading Recent Containers...'
+                        <Skeleton />
                       ) : (
                         <Section title="Recent Containers">
                           {recentContainersData &&