+++ packages/editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/editor-mobile-bridge
 
+## 6.7.10
+- [patch] [557a2b5734](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/557a2b5734):
+
+  - ED-5788: bump prosemirror-view and prosemirror-model
+
 ## 6.7.9
 - [patch] [01935688f8](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/01935688f8):